+++ templates/nasa_nex-gddp-cmip6_processing.py
@@ -4,7 +4,7 @@ from miranda import convert, io
 
 
 def main():
-    path_nex_out = Path("/jarre/scenario/jlavoie/data/NEX-GDDP-CMIP6_downloaded/")
+    path_nex_out = Path("INPUT_PATH")
     # put all nc that should be in 1 zarr together
     nex_files = convert.gather_nex(path_nex_out)
 
@@ -29,7 +29,7 @@ def main():
         io.write_dataset(
             ds,
             project="NEX-GDDP-CMIP6",
-            output_path=f"/jarre/scenario/jlavoie/data/NEX-GDDP-CMIP6_processed/NEX-GDDP-CMIP6/{end_path}",
+            output_path=f"OUTPUT_PATH/{end_path}",
             output_format="zarr",
             overwrite=True,
             compute=True,